re #94: don't accept "localhost" as hostname; read distribution id from /etc/lsb_release.
This is not a complete fix for this issue, since it doesn't work in a chroot,
and it doesn't support /usr/bin/lsb_release.

Oh, and did I mention how much the idea of having /usr/bin/lsb_release sucks?
On Debian that is a Python script. i.e. on every Avahi startup we'd have to
load an entire python interpreter into memory! That takes ages! I am not quite
sure if we really should pursue this bug any further because of this problem.
First, Debian should replace lsb_release by a saner implementation. In effect,
I will now mark this bug as WONTFIX.



git-svn-id: file:///home/lennart/svn/public/avahi/trunk@1379 941a03a8-eaeb-0310-b9a0-b1bbd8fe43fe