Revert "fix build on old freebsd, patch from zmi (Closes #158)"
This patch doesn't make sense, PCAP_D_IN is an enum field member
and thus #ifndef doesn't work.  This causes the build to break
on FreeBSD, despite the original intent to fix it.  They have
long shipped a patch to remove this.

I doubt we need to support such old pcap versions now, so just
removing this patch without finding some way to detect it better.

This reverts commit cf1f05215e3651ed47ea530edc7e924787de1bc6.