[Charles_Claggett@dell.com]
Add interface attribute for Avahi service files. If present,
the service or sub-type will publish only on the specified interface. If
not present the default behavior of publishing on all interfaces is unchanged.
The value for this attribute is the system ifname. If the interface is invalid,
the service is not published and an error is logged.