Adds support for filtering reflector advertisements This enhancement is provided courtesy of James Rudd <James.Rudd@sbhs.nsw.edu.au> and http://jrudd.org/2014/08/avahi-reflector-filtering-patch/
Allows Avahi-Daemon to  filter which advertisements are added to the cache to be reflected to different networks.
It checks incoming service names against a list defined  in avahi-daemon.conf [reflector] reflect-filters.
The list can be types of  services or can contain hostnames to match.
For example if we only allow AirPlay and AirTunes to be reflected between VLANs, so have  "_airplay._tcp.local,_raop._tcp.local" set.
For AirPrint set "_printer._tcp.local,_ipp._tcp.local,_pdl-datastream._tcp.local"
Remember to set firewall to permit traffic between LANs for the corresponding ports.

The patch will block the PTR and SRV advertisements but will still allow A records for machine name lookup.
All locally published services are still published even if they do not match the filter.

The filter also blocks local programs from seeing advertised programs so it is recommend to only enable it on a dedicated bonjour reflector server.