Use monotonic timer when possible
Use clock_gettime() + CLOCK_MONOTONIC as the clock source on platforms
that support it. This allows the various timers used within Avahi to
function correctly even when the system clock is changed. This is
especially important when the system clock is set backward by a
significant amount of time, as this can cause responses and queries in
their respective history queues to prevent future responses and queries
from being sent for an extended period of time.