filter denied virtual interfaces when adding addresses
netlink does not send an interface change notification when setting up a virtual
interface such as eth0:0: that information is only available in RTM_NEWADDR
messages.

It is therefore needed to check if a new address does belong to an allowed
interface label before actually adding it to the hardware interface.