Add support for binary values in TXT records in service files
This adds an attribute to the txt-record tag that describes the format
of the value part in the key=value part. For backwards-compatibility
it defaults to text if not specified. Other options include decoding
the value from hex encoding or from base64 encoding. That way we can
specify binary data in an XML file, that normally does not allow
arbitrary bytes in its contents.

These encodings result in larger data than the resulting binary data,
but that should not be a problem - TXT records ought to be small
anyway.