Remove default [rlimits] from avahi-daemon.conf
A number of issues with the default restrictive ulimits have recently
arisen.  Particularly the default data limit is being hit in some cases
and the process limit is being hit when multiple containers are in use
without UID mapping.

We remove all default rlimits and leave it up to the distribution to
impose some limit with systemd if they really desire to do so, but are
not recommended at this stage.

I also increase the example memory limits to double in case someone
wants to use them, as the existing limits are sometimes insufficient
(Closes: #113)

References:
https://bugs.launchpad.net/ubuntu/+source/avahi/+bug/1638345
https://bugs.launchpad.net/maas/+bug/1661869