Avahi: Protected strcmp against NULL pointers in multithreaded apps
A bug is triggered when a dbus message arrives, but the Avahi library
hasn't completed initialisation of a browser object (AvahiRecordBrowser
and potentially AvahiServiceTypeBrowser). Until initialisation has
completed, any dbus message that attempts to interrogate a browser
linked list (domain_browsers/service_type_browsers) will issue a segfault.

At various places in the code, b->path is involved in a string compare. By
initialising path to NULL, and adding it to the linked list immediately,
it is possible that strcmp uses a NULL ptr for its comparison (by code on
other threads) before the path is properly assigned by the avahi_strdup
call.

This fix checks for NULL before it is used.

Signed-off-by: Steven Goodwin <marquisdegeek@gmail.com>