avahi-python: Encode unicode strings as UTF-8
Previously, we would effectively encode anything representable in
Latin-1 as Latin-1, and crash on anything not representable in Latin-1:

>>> import avahi
>>> avahi.string_to_byte_array(u'©')
[dbus.Byte(169)]
>>> avahi.string_to_byte_array(u'\ufeff')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/lib/python2.7/dist-packages/avahi/__init__.py", line 94, in string_to_byte_array
    r.append(dbus.Byte(ord(c)))
ValueError: Integer outside range 0-255

This is particularly important for Python 3, where the str type
is a Unicode string.

The b'' syntax for bytestrings is supported since at least Python 2.7.

These functions now accept either Unicode strings (Python 2 unicode,
Python 3 str), which are encoded in UTF-8, or bytestrings
(Python 2 str, Python 3 bytes) which are taken as-is.

Signed-off-by: Simon McVittie <smcv@debian.org>