systemd: add basic modern security protection
Inspired by `systemd-resolved.service`, I went through the sandboxing
section of `man systemd.exec` and added settings accordingly.  A lot of
them are described as "recommended for long-running processes".

This applies a nice standard confinement to the service as a whole.
Most of these settings are just redundant protection for the filesystem.

I included a ban on debugging (ptrace), which an attacker might try using
to escape the chroot jail.

I banned user namespaces.  Just because we can, and they're a massive
feature, and there have been security bugs in the past.

I checked through our socket() calls, so I could set
RestrictAddressFamilies.  There's a *lot* of address families used for
different things.  Avahi does not need to use AF_BLUETOOTH for example, so
let's only allow the ones we need.

I *tried* setting NoNewPrivileges, and/or BoundingCapabilities based on
caps.c.  Unfortunately it's not that simple: they cause startup to fail.
So I haven't added either of them here.