avahi-autoipd: Initial IP selection based on MAC ignored first octet
When selecting the initial IP to start probing from, we multiply each
portion of the MAC by it's offset inside the MAC. The intent as per
RFC3927 is to try and generate the same IP from the same MAC and have it
be persistent across boots.

Unfortunately since the offset is 0-indexed, the first octet has no
effect on the result. In some cases, systems had a MAC on two interfaces
differing only in the locally administered bit which is in the first
octet and the same LLADDR was assigned to two interfaces on the same
host.

Change the calculation to offset the index by 1 so that all parts of the
MAC affect the calculation.

Upgrade Note: Unfortunately after upgrade all hosts will get a different
IP address calculated than they had under the old version. Anyone
assuming the IP will be the same will have to update to use the new
calculated IP.

Fixes #214