Listen on loopback interfaces
To allow services that advertise on the local machine, listen on
loopback interfaces even though they don't have the MULTICAST flag.

Currently transmitting on lo.IPv6 produces a sendmsg() network is
unreachable error, this appears to be "expected" and we should later try
to handle that better but for now you will see an error with --debug
enabled.

Closes #161
Closes #125