CI: turn on as many fatal warnings as possible
The idea is to turn on -Werror eventually to let compilers do their
thing and catch various mistakes automatically. At this point it isn't
possible to fully turn it on though so a bunch of warnings are just ignored.
They should be fixed but until then the CI should at least be able to prevent
new slip-ups from making it into the repository.