ci: import the OSS-Fuzz build script and fuzz targets
to make it easier to maintain them.

It's prompted by https://github.com/lathiat/avahi/pull/459
where the build script has to be edited to disable libsystemd.

This PR is based on https://github.com/google/oss-fuzz/pull/9107 and
includes all those changes as well:

* the fuzz targets are rewritten in C;

* the build script is changed to use the headers and static libraries
  from the the build directory

The license was discussed in
https://github.com/google/oss-fuzz/pull/10412#discussion_r1208682267.