docs: modernize INSTALL (#613)
* docs: modernize INSTALL

to make it actually work on Debian/Ubuntu. Other topics like turning
off D-Bus or running it on non-systemd systems should probably be
documented elsewhere eventually too.

The CI script is updated too to make sure the configure options
mentioned in INSTALL keep working on Ubuntu.

Addresses https://github.com/avahi/avahi/issues/612

* add SECURITY.md

pointing to the "Report a vulnerability" tab. People reporting
vulnerabilities are usually aware of that GitHub feature but it
shouldn't hurt to mention it explicitly just in case. It's also shown
when issues are opened.