core: reannounce after one second instead of withdrawing rrset (#654)
According to RFC6762, section 8.2, the host with the lexicographically
later record shall defer to the winning host by waiting one second and
then begin probing again.  If there is a conflict with a real host, this
host will have left its probing state after this time and will respond
to the conflicting probing packet.  So it should be sufficient to
withdraw the record in handle_conflict().

Closes: #654