configure.ac: Check if inotify functions are available without linking extra libs
On Linux the inotify API is a part of libc, so it is enough to check for the
<sys/inotify.h> header presence. On some OSes the inotify API is implemented as
a third party library called libinotify-kqueue. It requires correct CFLAGS and
LDFLAGS to properly link to it, so checking for the header only makes the build
system believe that inotify is available without any extra libs.

This change is required for a follow-up commit enabling the FreeBSD CI.