build: no longer carry dbm-related code
It already fails to compile everywhere so it doesn't make sense to keep
carrying it. The only known consumer switched from it in
https://github.com/OpenIndiana/oi-userland/pull/24599.

The dbm option is kept to prevent scripts disabling dbm explicitly with
--disable-dbm from breaking. --enable-dbm just fails loudly and says
that it's no longer supported. It's better than the current build
failures coming from the linker/makefiles (depending on where it fails).

Closes https://github.com/avahi/avahi/issues/738