ci: bring nss-mdns as well
avahi is shipped with nss-mdns sometimes so it makes sense to make sure
that they work fine together and it isn't completely broken. The CI there
is basically non-existent so it would make sense to double-check it just in case.

It also starts running some smoke tests on FreeBSD.

The following issues are skipped for now:
https://github.com/avahi/nss-mdns/issues/99
https://github.com/avahi/nss-mdns/issues/102
https://github.com/avahi/nss-mdns/issues/103
https://github.com/avahi/nss-mdns/issues/104
https://github.com/avahi/nss-mdns/issues/105