ci: trim distro checks
In some cases it doesn't matter much whether it's Ubuntu or FreeBSD.
What matters is whether the unit files are expected to be generated.
With this patch applied it should be easier to run the smoke tests
downstream outside of the upstream CI with its specific distros.

Also systemd-dev in installed explicitly because base images don't come
with it either.