ci: various tweaks
The cooldown option is set to prevent Dependabot from attempting to
switch to new releases as soon as they are available:
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#cooldown-.
It should prevent it from recommending releases that haven't been tested
yet.

The actions are pinned to their full-length commit SHAs. It's still the
most reliable way to make sure that tags aren't moved (immutable
releases are supposed to improve that but they aren't turned on
everywhere). CIFuzz and CFLite aren't pinned because they roll forward
by design.

The persist-credentials option is set to false because it's the most
sensible default. It got improved recently
(https://github.com/orgs/community/discussions/179107#discussioncomment-14906259)
but still. That being said all the actions are read-only mostly (apart
from a couple of places where security-events are writable to upload
SARIF analysis results) so it's a cosmetic change.