ci: use concurrency everywhere
differential-shellcheck is short-lived so it doesn't matter much whether
it's cancelled or not but it's still possible to force-push things fast
enough to get it to run in parallel and that's not desirable.

coverity is normally scheduled once a day so it doesn't matter much
whether it's cancelled properly or not but when it's tested in, say, PRs
it still makes sense to cancel it correctly.

zizmor is turned on in addition to actionlint to catch various things
when PRs are opened.