ci: run avahi-daemon under Valgrind everywhere
Up until this point avahi-daemon was run under Valgrind on Ubuntu only
because the output of Valgrind and its exit code were tracked only there.

On FreeBSD and Alpine the output of Valgrind is forwarded to /dev/null by
default (because of libdaemon forwarding stderr there in daemon_fork)
so Valgrind is run with --log-file instead and all the errors are looked
for there. The suppression file is added to suppress reports where
Valgrind complains about 4 file descriptors handled by libdaemon: stdin,
stdout, stderr and the syslog socket. It's needed on Ubuntu too but
Valgrind isn't new enough on Ubuntu 24.04 on the CI upstream so it
doesn't fail with --track-fds=yes yet (even though it complains about
the syslog fd).