build: use various sandboxing settings
to extend the builtin "sandbox" (where avahi switches to the avahi user,
calls chroot and drops all its privileges) with the seccomp filters and
so on on systems where they are available. Ideally avahi would ship a
sysusers.d config, the unit file would come with User= and so on but it
would require changing various things so as not to manually create the
runtime directory for example and that can be done later if necessary.
The goal for now is to ship a sensible unit file by default and not
conflict too much with some distros where for example sysusers.d configs
are already shipped.